Initial commit of forked code
This code is meant to find the latest file and copy it to a static image.  Note that when you first upload and run this, the latest file that you uploaded to your server will be displayed.  So that means it will try to show the PHP file and break the image until AmCrest refreshes with a new image to the server.